+++ glade_forms/lib/src/model/model.dart
@@ -1,1 +1,2 @@
+export 'async_glade_model.dart';
 export 'glade_model.dart';