+++ glade_forms/lib/src/widgets/async_glade_form_builder.dart
@@ -3,57 +3,64 @@ import 'package:glade_forms/src/widgets/glade_model_provider.dart';
 import 'package:provider/provider.dart';
 
 import '../model/glade_model_base.dart';
+import 'glade_form_builder.dart';
 
-typedef GladeFormWidgetBuilder<M extends GladeModelBase> = Widget Function(
-  BuildContext context,
-  M model,
-  Widget? child,
-);
-
-class GladeFormBuilder<M extends GladeModelBase> extends StatelessWidget {
+class AsyncGladeFormBuilder<M extends GladeModelBase> extends StatelessWidget {
   // ignore: prefer-correct-callback-field-name, ok name
   final CreateModelFunction<M>? create;
   final M? value;
   // ignore: prefer-correct-callback-field-name, ok name
   final GladeFormWidgetBuilder<M> builder;
+  final WidgetBuilder? initializeBuilder;
   final Widget? child;
 
-  factory GladeFormBuilder({
+  factory AsyncGladeFormBuilder({
     required GladeFormWidgetBuilder<M> builder,
+    WidgetBuilder? initializeBuilder,
     Key? key,
     Widget? child,
   }) =>
-      GladeFormBuilder._(builder: builder, key: key, child: child);
+      AsyncGladeFormBuilder._(
+        builder: builder,
+        key: key,
+        child: child,
+        initializeBuilder: initializeBuilder,
+      );
 
-  const GladeFormBuilder._({
+  const AsyncGladeFormBuilder._({
     required this.builder,
+    required this.initializeBuilder,
     super.key,
     this.create,
     this.value,
     this.child,
   });
 
-  factory GladeFormBuilder.create({
+  factory AsyncGladeFormBuilder.create({
     required CreateModelFunction<M> create,
     required GladeFormWidgetBuilder<M> builder,
+    WidgetBuilder? initializeBuilder,
     Widget? child,
     Key? key,
   }) =>
-      GladeFormBuilder._(
+      AsyncGladeFormBuilder._(
         builder: builder,
+        initializeBuilder: initializeBuilder,
         create: create,
         key: key,
         child: child,
       );
 
-  factory GladeFormBuilder.value({
+  factory AsyncGladeFormBuilder.value({
     required GladeFormWidgetBuilder<M> builder,
     required M value,
+    WidgetBuilder? initializeBuilder,
     Widget? child,
     Key? key,
   }) =>
-      GladeFormBuilder._(
+      AsyncGladeFormBuilder._(
         builder: builder,
+        initializeBuilder: initializeBuilder,
         value: value,
         key: key,
         child: child,
@@ -61,18 +68,24 @@ class GladeFormBuilder<M extends GladeModelBase> extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final consumerBuilder = (context, model, child) {
+      if (model.isInitialized) return builder(context, model, child);
+
+      return initializeBuilder?.call(context) ?? CircularProgressIndicator();
+    };
+
     if (create case final createFn?) {
       return GladeModelProvider(
         create: createFn,
-        child: Consumer<M>(builder: builder, child: child),
+        child: Consumer<M>(builder: consumerBuilder, child: child),
       );
     } else if (value case final modelValue?) {
       return GladeModelProvider.value(
         value: modelValue,
-        child: Consumer<M>(builder: builder, child: child),
+        child: Consumer<M>(builder: consumerBuilder, child: child),
       );
     }
 
-    return Consumer<M>(builder: builder, child: child);
+    return Consumer<M>(builder: consumerBuilder, child: child);
   }
 }