+++ glade_forms/lib/src/widgets/widgets.dart
@@ -1,3 +1,4 @@
+export 'async_glade_form_builder.dart';
 export 'glade_form_builder.dart';
 export 'glade_form_consumer.dart';
 export 'glade_form_listener.dart';