+++ storybook/lib/usecases/dependencies/checkbox_dependency_change.dart
@@ -19,7 +19,7 @@ class AgeRestrictedModel extends GladeModel {
   List<GladeInput<Object?>> get inputs => [nameInput, ageInput, vipInput];
 
   @override
-  void initialize() {
+  void initializeInputs() {
     nameInput = GladeInput.stringInput(
       inputKey: 'name-input',
       defaultTranslations: DefaultTranslations(
@@ -69,8 +69,6 @@ class AgeRestrictedModel extends GladeModel {
         }
       },
     );
-
-    super.initialize();
   }
 }