+++ storybook/lib/usecases/onchange/two_way_checkbox_change.dart
@@ -19,7 +19,7 @@ class AgeRestrictedModel extends GladeModel {
   List<GladeInput<Object?>> get inputs => [nameInput, ageInput, vipInput];
 
   @override
-  void initialize() {
+  void initializeInputs() {
     nameInput = GladeInput.stringInput(
       inputKey: 'name-input',
       defaultTranslations: DefaultTranslations(
@@ -67,8 +67,6 @@ class AgeRestrictedModel extends GladeModel {
         }
       },
     );
-
-    super.initialize();
   }
 }